Skriver om vurderingArtikkel16Vedtak til typescript, og skiller ut separate komponenter i egne filer. Trakk også inn det jeg kunne i useSelectors
