+++ src/ducks/lovvalgsperioder/selectors.js
@@ -58,6 +58,12 @@ export const InnvilgelsesResultatSelector = createSelector(
   LovvalgsperiodeSelector,
   (lovvalgsperiode) => lovvalgsperiode.innvilgelsesResultat
 );
-export const FomDatoSelector = createSelector(LovvalgsperiodeSelector, (lovvalgsperiode) => lovvalgsperiode.fomDato);
+export const FomDatoSelector = createSelector(
+  (state) => LovvalgsperiodeSelector(state),
+  (lovvalgsperiode) => lovvalgsperiode.fomDato
+);
 
-export const TomDatoSelector = createSelector(LovvalgsperiodeSelector, (lovvalgsperiode) => lovvalgsperiode.tomDato);
+export const TomDatoSelector = createSelector(
+  (state) => LovvalgsperiodeSelector(state),
+  (lovvalgsperiode) => lovvalgsperiode.tomDato
+);
