+++ src/felleskomponenter/stegvelger/Stegvelger.jsx
@@ -417,7 +417,6 @@ class Stegvelger extends Component {
       erArbeidEttLand: props.erArbeidEttLand,
       maritimtarbeid: props.maritimtarbeid,
       hjemmebaser: props.hjemmebaser,
-      harValgtNorskArbeidsgiver: props.harValgtNorskArbeidsgiver,
       medfolgendeBarn: props.medfolgendeBarn,
       lagredeVirksomheter: props.lagredeVirksomheter,
       bestemmelser: props.bestemmelser,
@@ -622,7 +621,6 @@ Stegvelger.propTypes = {
   erArbeidEttLand: PT.bool.isRequired,
   videresend: PT.func.isRequired,
   bestillAnmodningsperioder: PT.func.isRequired,
-  harValgtNorskArbeidsgiver: PT.bool.isRequired,
   medfolgendeBarn: PT.array.isRequired,
   lagredeVirksomheter: PT.array.isRequired,
   sakstype: PT.string,
@@ -722,7 +720,6 @@ const mapStateToProps = (state) => ({
   maritimtarbeid: formSelectors.MaritimtArbeidSelector(state),
   hjemmebaser: mottatteOpplysningerSelectors.HjemmebaserSelector(state),
   erArbeidEttLand: behandlingerSelectors.ErArbeidEttLand(state),
-  harValgtNorskArbeidsgiver: flytSelectors.HarValgtNorskArbeidsgiverSelector(state),
   medfolgendeBarn: mottatteOpplysningerSelectors.MedfolgendeBarnSelector(state),
   lagredeVirksomheter: oppsummertfaktaSelectors.VirksomhetIDerSelector(state),
   soknadsperiode: mottatteOpplysningerSelectors.PeriodeSelector(state),
